requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
django-environ>=0.11,<1.0
gunicorn>=21.2,<22.0
Pillow>=10.0.0
requests>=2.31.0
google-generativeai>=0.3.0
